requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
httpx==0.25.2
pandas==2.1.3
python-dotenv==1.0.0
numpy==1.25.2
dagster==1.11.8
dagster-postgres==0.27.8
dagster-webserver==1.11.8
